Simplify by putting all scans in core-scanner. This deletes libs/page-scanners and retires libs/solutions-scanner..
