+++ apps/scan-engine/src/app.module.ts
@@ -6,7 +6,7 @@ import { DatabaseModule } from '@app/database';
 import { MessageQueueModule } from '@app/message-queue';
 import { QueueModule } from '@app/queue';
 
-import { SolutionsScannerModule } from 'libs/solutions-scanner/src';
+import { CoreScannerModule } from 'libs/core-scanner/src';
 import { ScanEngineConsumer } from './scan-engine.consumer';
 
 const logger = pino(
@@ -27,7 +27,7 @@ setInterval(function () {
   imports: [
     MessageQueueModule,
     DatabaseModule,
-    SolutionsScannerModule,
+    CoreScannerModule,
     LoggerModule.forRoot(),
     QueueModule,
   ],
