+++ libs/core-scanner/src/pages/helpers.ts
