+++ libs/core-scanner/src/pages/home-page/index.ts
@@ -1,13 +1,10 @@
 import { Logger } from '@nestjs/common';
 
-import { SolutionsInputDto } from '@app/solutions-scanner/solutions.input.dto';
+import { CoreInputDto } from '@app/core-scanner/core.input.dto';
 
 import { getHttpsUrl } from '../helpers';
 
-export const createHomePageScanner = (
-  logger: Logger,
-  input: SolutionsInputDto,
-) => {
+export const createHomePageScanner = (logger: Logger, input: CoreInputDto) => {
   const url = getHttpsUrl(input.url);
   /*const response = await page.goto(url, {
     waitUntil: 'networkidle2',
