+++ libs/core-scanner/src/pages/index.ts
