+++ libs/core-scanner/src/pages/sitemap-xml.ts
@@ -1,12 +1,12 @@
 import { Logger } from '@nestjs/common';
 import { Page, Response } from 'puppeteer';
 
-import { SolutionsInputDto } from '@app/solutions-scanner/solutions.input.dto';
+import { CoreInputDto } from '@app/core-scanner/core.input.dto';
 import { getHttpsUrl, getMIMEType } from './helpers';
 
 export const createSitemapXmlScanner = (
   logger: Logger,
-  input: SolutionsInputDto,
+  input: CoreInputDto,
 ) => {
   const url = getHttpsUrl(input.url);
   return async (sitemapPage: Page) => {
