+++ libs/core-scanner/src/scans/core.test.ts
