+++ libs/core-scanner/src/scans/seo.test.ts
