+++ libs/core-scanner/src/scans/seo.ts
