+++ libs/core-scanner/src/scans/third-party.test.ts
