+++ libs/core-scanner/src/scans/third-party.ts
