--- libs/solutions-scanner/src/index.ts
@@ -1,2 +1,0 @@
-export * from './solutions-scanner.module';
-export * from './solutions-scanner.service';
