--- libs/solutions-scanner/src/solutions-scanner.module.ts
@@ -1,11 +1,0 @@
-import { HttpModule, Module } from '@nestjs/common';
-
-import { BrowserModule } from '@app/browser';
-import { SolutionsScannerService } from './solutions-scanner.service';
-
-@Module({
-  imports: [BrowserModule, HttpModule],
-  providers: [SolutionsScannerService],
-  exports: [SolutionsScannerService],
-})
-export class SolutionsScannerModule {}
