--- libs/solutions-scanner/src/testPageSource.ts
@@ -1,5812 +1,0 @@
-const source = `
-<!DOCTYPE html>
-<!--[if lt IE 9]><html class="lt-ie9"><![endif]-->
-<!--[if gt IE 8]><!--><html lang="en"><!--<![endif]-->
-
-  
-
-<!-- site.url: https://18f.gsa.gov -->
-<!-- site.baseurl:  -->
-<!-- siteurl: https://18f.gsa.gov -->
-
-<head>
-  
-<!-- Basic Page Needs
-================================================== -->
-  <meta charset="utf-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="google-site-verification" content="vmVo2wbQpKGMHVpllQP7bQ0w5gICWD356VH9DNbv5f4" />
-
-<!-- Mobile Specific Metas
-================================================== -->
-  <meta name="HandheldFriendly" content="True">
-  <meta name="MobileOptimized" content="320">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-<!-- Meta tags
-================================================== -->
-
-
-<!-- Creator -->
-<meta name="twitter:creator" content="@18F">
-<!-- Title -->
-
-  <title>18F: Digital service delivery | Home</title>
-  <meta property="og:title" content="18F: Digital service delivery | Home" />
-  <meta name="twitter:title" content="18F: Digital service delivery | Home">
-
-
-<!-- Twitter Site -->
-<meta name="twitter:site" content="@18F" />
-
-<!-- Description -->
-
-  <meta name="description" content="18F builds effective, user-centric digital services focused on the interaction between government and the people and businesses it serves.">
-  <meta property="og:description" content="18F builds effective, user-centric digital services focused on the interaction between government and the people and businesses it serves." />
-  <meta property="twitter:description" content="18F builds effective, user-centric digital services focused on the interaction between government and the people and businesses it serves." />
-
-
-<!-- Image -->
-
-  <meta name="twitter:card" content="summary">
-  <meta name="twitter:image" content="https://18f.gsa.gov/assets/img/logos/18F-Logo-M.png">
-  <meta property="og:image" content="https://18f.gsa.gov/assets/img/logos/18F-Logo-M.png">
-
-
-<!-- Author -->
-
-
-<!-- Date -->
-
-
-<!-- Tags -->
-
-
-  <link rel="canonical" href="https://18f.gsa.gov/" />
-
-<!--[if lt IE 9]>
-  <script src="/assets/js/vendor/html5shiv.js"></script>
-  <script src="/assets/js/vendor/respond.js"></script>
-  <script src="/assets/js/vendor/selectivizr-min.js"></script>
-<![endif]-->
-
-<!-- Favicons
-================================================== -->
-
-  <!--  Touch icon for iOS 2.0+ and Android 2.1+ -->
-  <link rel="apple-touch-icon-precomposed" href="/assets/img/favicons/18F-Logo-152.png">
-
-  <!--  Desktop icon -->
-  <link rel="icon" href="/assets/img/favicons/18F-Logo-152.png" type="image/png">
-
-  <!--  More info: https://github.com/audreyr/favicon-cheat-sheet -->
-
-<!-- CSS
-=================================================== -->
-  <link rel="stylesheet" href="/assets/css/main.css" media="all" />
-  <link rel="stylesheet" href="/assets/css/print.css" media="print" />
-
-</head>
-
-
-  <body class="page-home
-   ">
-
-    
-      <a class="skipnav" href="#main-content">Skip to main content</a>
-<section class="usa-banner" aria-label="Official government website">
-  <div class="usa-accordion">
-    <header class="usa-banner__header">
-      <div class="usa-banner__inner">
-        <div class="grid-col-auto">
-          <img
-            class="usa-banner__header-flag"
-            src="/assets/img/us_flag_small.png"
-            alt="U.S. flag"
-          >
-        </div>
-        <div class="grid-col-fill tablet:grid-col-auto">
-          <p class="usa-banner__header-text">An official website of the United States government</p>
-          <p class="usa-banner__header-action" aria-hidden="true">Here’s how you know</p>
-        </div>
-        <button
-          class="usa-accordion__button usa-banner__button"
-          aria-expanded="false"
-          aria-controls="gov-banner"
-        >
-          <span class="usa-banner__button-text">
-            Here’s how you know
-          </span>
-        </button>
-      </div>
-    </header>
-    <div class="usa-banner__content usa-accordion__content" id="gov-banner">
-      <div class="grid-row grid-gap-lg">
-        <div class="usa-banner__guidance tablet:grid-col-6">
-          <img
-            class="usa-banner__icon usa-media-block__img"
-            src="/assets/img/icon-dot-gov.svg"
-            role="img"
-            alt="Dot gov"
-          >
-          <div class="usa-media-block__body">
-            <p>
-              <strong>
-                Official websites use .gov
-              </strong>
-              <br/>
-              A <strong>.gov</strong> website belongs to an official government organization in the United States.
-            </p>
-          </div>
-        </div>
-        <div class="usa-banner__guidance tablet:grid-col-6">
-          <img
-            class="usa-banner__icon usa-media-block__img"
-            src="/assets/img/icon-https.svg"
-            role="img"
-            alt="Https"
-          >
-          <div class="usa-media-block__body">
-            <p>
-              <strong>
-                Secure .gov websites use HTTPS
-              </strong>
-              <br/>
-              A <strong>lock</strong> (
-                <span class="icon-lock">
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="52"
-                    height="64"
-                    viewBox="0 0 52 64"
-                    class="usa-banner__lock-image"
-                    role="img"
-                    aria-labelledby="banner-lock-title banner-lock-description"
-                  >
-                    <title id="banner-lock-title">Lock</title>
-                    <desc id="banner-lock-description">A locked padlock</desc>
-                    <path
-                      fill="#000000"
-                      fill-rule="evenodd"
-                      d="M26 0c10.493 0 19 8.507 19 19v9h3a4 4 0 0 1 4 4v28a4 4 0 0 1-4 4H4a4 4 0 0 1-4-4V32a4 4 0 0 1 4-4h3v-9C7 8.507 15.507 0 26 0zm0 8c-5.979 0-10.843 4.77-10.996 10.712L15 19v9h22v-9c0-6.075-4.925-11-11-11z"
-                    />
-                  </svg>
-                </span>
-                ) or <strong>https://</strong> means you’ve safely connected to the .gov website. Share sensitive information only on official, secure websites.
-            </p>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</section>
-
-<div class="overlay"></div>
-<header class="usa-header usa-header--basic">
-  <div class="usa-nav-container">
-    <div class="usa-navbar desktop:maxw-card height-auto flex-align-center">
-      <div class="usa-logo">
-        <a class="site-logo media_link" href="/">
-          <img src="/assets/img/logos/18f-logo.svg" alt="18F
-          logo"  aria-hidden="false"  ></a
-        >
-        
-        <h1 class="usa-sr-only">18F</h1>
-        
-      </div>
-      <button class="usa-button--accent-cool usa-menu-btn">Menu</button>
-    </div>
-    
-
-
-  
-  <!-- prettier-ignore -->
-<nav aria-label="Primary navigation" class="usa-nav">
-  <button class="usa-nav__close">
-    <img src="/assets/img/close.svg" alt="close">
-  </button>
-  <ul class="usa-nav__primary usa-accordion">
-    
-    <li class="usa-nav__primary-item">
-        
-
-        
-
-        <a href="/what-we-deliver/" ><span>What we deliver</span></a>
-      </li>
-    
-    <li class="usa-nav__primary-item">
-        
-
-        
-
-        <a href="/how-we-work/" ><span>How we work</span></a>
-      </li>
-    
-    <li class="usa-nav__primary-item">
-        
-
-        
-
-        <a href="/guides/" ><span>Guides</span></a>
-      </li>
-    
-    <li class="usa-nav__primary-item">
-        
-
-        
-
-        <a href="/blog/" ><span>Blog</span></a>
-      </li>
-    
-    <li class="usa-nav__primary-item">
-        
-
-        
-
-        <a href="/contact/" ><span>Contact</span></a>
-      </li>
-    
-  </ul>
-  <form
-    class="usa-search usa-search--small"
-    role="search"
-    id="search_form"
-    action="https://search.usa.gov/search"
-    accept-charset="UTF-8"
-    method="get"
-  >
-    <input name="utf8" type="hidden" value="&#x2713;" />
-    <label class="usa-sr-only" for="search-field-small">Search</label>
-    <input class="usa-input" type="hidden" name="affiliate" id="affiliate" value="18f-site" >
-    <input class="usa-input" type="search" name="query" id="query" autocomplete="off">
-    <button class="usa-button" type="submit">
-      <span class="usa-sr-only">Search</span>
-    </button>
-  </form>
-</nav>
-
-
-
-
-
-
-
-
-  </div>
-</header>
-
-    
-
-    <div class="main-content" id="main-content">
-      <section class="background-dark">
-  <section class="usa-hero usa-hero--primary">
-    <div class="grid-container">
-      <div
-        class="usa-hero__callout background-dark tablet:grid-col-8"
-      >
-        <h2>
-          18F partners with agencies to improve the user experience of
-          government services by helping them build and buy technology.
-        </h2>
-        <a href="/contact/" tabindex="-1"
-          ><button class="usa-button usa-button-secondary usa-button--big">
-            Get in touch
-          </button></a
-        >
-      </div>
-    </div>
-  </section>
-</section>
-
-<section class="usa-section grid-container">
-  <section class="pad-right-left">
-    <div class="home-testimonial">
-      We engaged with 18F, and it was a game-changer for us. They came in and
-      helped the IT team and business owners get involved in the agile
-      development process. It changed our whole methodology.
-      <span>- Alec Palmer, FEC’s Chief Information Officer</span>
-    </div>
-  </section>
-</section>
-
-<section class="usa-section grid-container">
-  <section class="grid-gap">
-    <h2>We support your mission</h2>
-    <section class="tablet:grid-col-8">
-      <h3 class="p-line-height p-normal pad-bottom-mobile">
-        As an office within the General Services Administration, we know how to
-        work with government. Federal, state, and local governments can partner
-        with 18F to deliver projects that fulfill their mission, stay within
-        budget, and use leading technology practices.
-      </h3>
-    </section>
-  </section>
-</section>
-
-<section class="usa-section background-gray">
-  <section class="grid-container">
-     
-    <h2>Some agencies we’ve worked with</h2>
-    <ul class="grid-row list-columns grid-gap">
-      
-      <li class="tablet:grid-col-4">
-        <ul class="list-columns list-images">
-          
-          <li class="list-images-item">
-            <img
-              class="list-images-image"
-              src="/assets/img/logos/agencies/USDA.svg"
-              alt="/assets/img/logos/agencies/USDA.svg logo"
-            />
-            
-            <span class="list-images-text">Department of Agriculture</span>
-            
-          </li>
-          
-          <li class="list-images-item">
-            <img
-              class="list-images-image"
-              src="/assets/img/logos/agencies/usmc.png"
-              alt="/assets/img/logos/agencies/usmc.png logo"
-            />
-            
-            <span class="list-images-text">U.S. Marine Corps</span>
-            
-          </li>
-          
-        </ul>
-      </li>
-      
-      <li class="tablet:grid-col-4">
-        <ul class="list-columns list-images">
-          
-          <li class="list-images-item">
-            <img
-              class="list-images-image"
-              src="/assets/img/logos/agencies/OMB.svg"
-              alt="/assets/img/logos/agencies/OMB.svg logo"
-            />
-            
-            <span class="list-images-text">Office of Management and Budget</span>
-            
-          </li>
-          
-          <li class="list-images-item">
-            <img
-              class="list-images-image"
-              src="/assets/img/logos/agencies/centers-for-medicare-and-medicaid.png"
-              alt="/assets/img/logos/agencies/centers-for-medicare-and-medicaid.png logo"
-            />
-            
-            <span class="list-images-text">Center for Medicaid and Medicare Services</span>
-            
-          </li>
-          
-        </ul>
-      </li>
-      
-      <li class="tablet:grid-col-4">
-        <ul class="list-columns list-images">
-          
-          <li class="list-images-item">
-            <img
-              class="list-images-image"
-              src="/assets/img/logos/agencies/usaf.png"
-              alt="/assets/img/logos/agencies/usaf.png logo"
-            />
-            
-            <span class="list-images-text">U.S. Air Force</span>
-            
-          </li>
-          
-          <li class="list-images-item">
-            <img
-              class="list-images-image"
-              src="/assets/img/logos/agencies/nga.png"
-              alt="/assets/img/logos/agencies/nga.png logo"
-            />
-            
-            <span class="list-images-text">National Geospatial-Intelligence Agency</span>
-            
-          </li>
-          
-        </ul>
-      </li>
-      
-    </ul>
-    <p>
-      <a
-        class="link-arrow-right post-link-continue_reading"
-        href="/what-we-deliver/#some-agencies-weve-worked-with"
-      >
-        See more agency partners <svg class="icon" xmlns="http://www.w3.org/2000/svg" width="444.819" height="444.819" viewBox="0 0 444.819 444.819" role="img" aria-hidden="true">
-  <title>Arrow right</title>
-  <desc>Arrow pointing to the right</desc>
-  <path fill="#046b99" d="M352.025 196.712L165.885 10.848C159.028 3.615 150.468 0 140.185 0s-18.84 3.62-25.696 10.848l-21.7 21.416c-7.045 7.043-10.567 15.604-10.567 25.692 0 9.897 3.52 18.56 10.566 25.98L231.544 222.41 92.785 361.168c-7.04 7.043-10.563 15.604-10.563 25.693 0 9.9 3.52 18.566 10.564 25.98l21.7 21.417c7.043 7.043 15.612 10.564 25.697 10.564 10.09 0 18.656-3.52 25.697-10.564L352.025 248.39c7.046-7.423 10.57-16.084 10.57-25.98.002-10.09-3.524-18.655-10.57-25.698z"/>
-</svg>
-
-      </a>
-    </p>
-  </section>
-</section>
-
-<section class="usa-section">
-  <section class="grid-container">
-    <h2>What 18F can do</h2>
-    <div class="grid-col-12">
-      <ul class="grid-row list-columns grid-gap">
-        <li class="tablet:grid-col-4">
-          <div class="icon-list">
-            <figure class="icon-list-image">
-              <?xml version="1.0" encoding="UTF-8"?>
-<svg width="97px" height="97px" viewBox="0 0 97 97" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" role="img" aria-hidden="true">
-    <!-- Generator: Sketch 40.3 (33839) - http://www.bohemiancoding.com/sketch -->
-    <title>User centered-design</title>
-    <desc>Demonstrating our commitment to our users</desc>
-    <defs></defs>
-    <g stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
-        <g transform="translate(-373.000000, -874.000000)">
-            <path d="M404.8772,936.0624 C387.9244,938.3756 375.8552,944.2024 375.8552,951.052 C375.8552,959.9316 396.068,967.1316 421.002,967.1316 C445.936,967.1316 466.1488,959.9316 466.1488,951.052 C466.1488,944.2692 454.3156,938.4936 437.6296,936.1352" id="Shape1" stroke="#00CFFF" stroke-width="4"></path>
-            <path d="M421.0024,955.1044 L421.0032,925.4184" id="Shape2" stroke="#00CFFF" stroke-width="4"></path>
-            <path d="M425.938,881.8972 C425.938,884.6224 423.7288,888.4284 421.0032,888.4284 C418.278,888.4284 416.0684,884.6228 416.0684,881.8972 C416.0684,879.1716 418.2776,876.9624 421.0032,876.9624 C423.7288,876.9628 425.938,879.172 425.938,881.8972 L425.938,881.8972 Z" id="Shape3" stroke="#00CFFF" stroke-width="4"></path>
-            <path d="M414.6168,955.1044 L412.6668,920.002 C412.638,919.2928 410.3608,920.216 410.3356,919.5436 C409.408,894.7168 411.472,893.138 411.7052,892.9428 C411.9796,892.866 414.232,892.8052 414.906,892.7628 C415.8172,892.7056 416.6028,893.262 417.0212,894.0736 C418.5052,896.9516 423.5008,896.9516 424.9848,894.0736 C425.4032,893.262 426.1888,892.7056 427.1,892.7628 C427.774,892.8052 430.0264,892.866 430.3008,892.9428 C430.534,893.138 432.5984,894.7168 431.6708,919.544 C431.6456,920.2164 429.3684,919.2932 429.3396,920.0024 L427.3876,955.1048" id="Shape4" stroke="#00CFFF" stroke-width="4"></path>
-            <rect x="373.002" y="874.0016" width="96" height="96"></rect>
-        </g>
-    </g>
-</svg>
-
-            </figure>
-            <div class="icon-list-text">
-              <p class="p-bold">
-                Modernize software development processes while introducing good
-                agile and human-centered design practices to your agency
-              </p>
-            </div>
-          </div>
-        </li>
-        <li class="tablet:grid-col-4">
-          <div class="icon-list">
-            <figure class="icon-list-image">
-              <?xml version="1.0" encoding="UTF-8"?>
-<svg width="60px" height="60px" viewBox="0 0 60 60" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" role="img" aria-hidden="true">
-    <!-- Generator: Sketch 40.3 (33839) - http://www.bohemiancoding.com/sketch -->
-    <title>Monitor icon</title>
-    <desc>Improve public facing interfaces</desc>
-    <defs></defs>
-    <g stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
-        <g transform="translate(-540.000000, -542.000000)" stroke="#00CFFF" stroke-width="2">
-            <g transform="translate(541.000000, 543.000000)">
-                <path d="M33.4119424,50 C38.1582333,50 39.249395,53.6406576 40.7129472,56.6090012 C40.7129472,57.1041006 40.2215118,57.5090909 39.6160333,57.5090909 L18.0969139,57.5090909 C17.4914354,57.5090909 17,57.0189976 17,56.5238982 C18.5045677,53.5938509 19.6672564,50 24.3002545,50 L33.4119424,50 L33.4119424,50 Z" id="Shape"></path>
-                <path d="M53.504688,0 L4.00876256,0 C1.73165338,0 0,1.93534303 0,4.21335091 L0,36.2263576 C0,38.505617 1.73165338,40.0484848 4.00876256,40.0484848 L53.504688,40.0484848 C55.7842981,40.0484848 57.5217037,38.505617 57.5217037,36.2263576 L57.5217037,4.21335091 C57.5217037,1.93534303 55.7842981,0 53.504688,0 L53.504688,0 Z"></path>
-                <path d="M28,50.00556 L28,45"></path>
-
-                <path d="M0.5,7.18 L56.5596315,7.18"></path>
-                <rect x="6" y="26" width="7.25836683" height="7.25878788"></rect>
-                <rect x="6" y="13" width="7.25836683" height="7.25878788"></rect>
-                <path d="M18,14 L34.1667419,14"></path>
-                <path d="M18,27 L34.1667419,27"></path>
-                <path d="M18,18 L25.0349906,18"></path>
-                <path d="M18,31 L25.0349906,31"></path>
-            </g>
-        </g>
-    </g>
-</svg>
-
-            </figure>
-            <div class="icon-list-text">
-              <p class="p-bold">
-                Improve public-facing services like websites or applications
-              </p>
-            </div>
-          </div>
-        </li>
-        <li class="tablet:grid-col-4">
-          <div class="icon-list">
-            <figure class="icon-list-image">
-              <?xml version="1.0" encoding="UTF-8"?>
-<svg width="56px" height="59px" viewBox="0 0 56 59" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" role="img" aria-hidden="true">
-    <!-- Generator: Sketch 40.3 (33839) - http://www.bohemiancoding.com/sketch -->
-    <title>Folder with clock</title>
-    <desc>Representative of design systems</desc>
-    <defs></defs>
-    <g stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
-        <g transform="translate(-541.000000, -803.000000)">
-            <g transform="translate(540.000000, 804.000000)">
-                <path d="M12.8566853,7.21007066 C12.2761355,5.85853777 11.3766285,4 10.6746846,4 L3.81758016,4 C2.81548319,4 2,4.81019956 2,5.80692506 L2,40.4729035 C2,41.4686687 2.81524516,42.2781481 3.81758016,42.2781481 L16.0564928,42.2781481" stroke="#00CFFF" stroke-width="2"></path>
-                <path d="M47.2179998,18.5191509 L47.2179998,9.83381168 C47.2179998,8.82220251 46.4027547,8 45.4013718,8 L41,8" stroke="#00CFFF" stroke-width="2"></path>
-                <path d="M47.0762737,16 L2,16" stroke="#00CFFF" stroke-width="2"></path>
-                <polygon stroke="#00CFFF" stroke-width="2" points="42.18665 15.2590321 7.2353642 15.2590321 6 7.8496071 39.5005065 1"></polygon>
-                <ellipse stroke="#00CFFF" stroke-width="2" cx="38.6057704" cy="39.747485" rx="16.6057704" ry="16.747485"></ellipse>
-                <polyline stroke="#00CFFF" stroke-width="2" points="46.7827492 39.5992633 38 39.5992633 38 27"></polyline>
-            </g>
-        </g>
-    </g>
-</svg>
-
-            </figure>
-            <div class="icon-list-text">
-              <p class="p-bold">
-                Digitize and streamline internal systems to save time and
-                increase accuracy
-              </p>
-            </div>
-          </div>
-        </li>
-      </ul>
-    </div>
-  </section>
-</section>
-
-<section class="usa-section background-gray">
-  <section class="usa-section grid-container">
-    <h2>Featured projects</h2>
-    <div class="grid-row grid-gap">
-      
-
-
-
-
-
-
-
-
-<article class="card tablet:grid-col-4 margin-bottom-4">
-  <div class="card-link" data-href="/what-we-deliver/fec-gov/" tabindex="-1">
-    <div role="img"
-         title="Screenshot of the FEC data explorer with stylized magnifying glass."
-         class="card-image-bg "
-         
-         style="background-image: url(/assets/img/home/hero-fec.png);">
-      
-    </div>
-    <div class="card-banner">
-      
-        <p class="card-tag">Federal Election Commission</p>
-      
-      
-        <a href="/what-we-deliver/fec-gov/" tabindex="-1" class="card-link-tagline"><h3 class="h4">Easy-to-use campaign data</h3></a>
-      
-      <p class="card-description">Rethinking how the Federal Election Commission provides data, instructions, and legal resources to the public.</p>
-    </div>
-    <a href="/what-we-deliver/fec-gov/" class="card-cta link-arrow-right" tabindex="0">
-      Read more
-      <span class="usa-sr-only">about Easy-to-use campaign data</span>
-      <svg class="icon" xmlns="http://www.w3.org/2000/svg" width="444.819" height="444.819" viewBox="0 0 444.819 444.819" role="img" aria-hidden="true">
-  <title>Arrow right</title>
-  <desc>Arrow pointing to the right</desc>
-  <path fill="#046b99" d="M352.025 196.712L165.885 10.848C159.028 3.615 150.468 0 140.185 0s-18.84 3.62-25.696 10.848l-21.7 21.416c-7.045 7.043-10.567 15.604-10.567 25.692 0 9.897 3.52 18.56 10.566 25.98L231.544 222.41 92.785 361.168c-7.04 7.043-10.563 15.604-10.563 25.693 0 9.9 3.52 18.566 10.564 25.98l21.7 21.417c7.043 7.043 15.612 10.564 25.697 10.564 10.09 0 18.656-3.52 25.697-10.564L352.025 248.39c7.046-7.423 10.57-16.084 10.57-25.98.002-10.09-3.524-18.655-10.57-25.698z"/>
-</svg>
-
-    </a>
-  </div>
-</article>
-
- 
-
-
-
-
-
-
-
-
-<article class="card tablet:grid-col-4 margin-bottom-4">
-  <div class="card-link" data-href="/what-we-deliver/data-act/" tabindex="-1">
-    <div role="img"
-         title="Data Act logo alongside the words Better Data"
-         class="card-image-bg "
-         
-         style="background-image: url(/assets/img/projects/data-act-better-data.jpg);">
-      
-    </div>
-    <div class="card-banner">
-      
-        <p class="card-tag">Department of the Treasury</p>
-      
-      
-        <a href="/what-we-deliver/data-act/" tabindex="-1" class="card-link-tagline"><h3 class="h4">Implementing a government-wide law</h3></a>
-      
-      <p class="card-description">Implementing the Digital Accountability and Transparency Act with one of the largest government-wide agile projects ever undertaken.</p>
-    </div>
-    <a href="/what-we-deliver/data-act/" class="card-cta link-arrow-right" tabindex="0">
-      Read more
-      <span class="usa-sr-only">about Implementing a government-wide law</span>
-      <svg class="icon" xmlns="http://www.w3.org/2000/svg" width="444.819" height="444.819" viewBox="0 0 444.819 444.819" role="img" aria-hidden="true">
-  <title>Arrow right</title>
-  <desc>Arrow pointing to the right</desc>
-  <path fill="#046b99" d="M352.025 196.712L165.885 10.848C159.028 3.615 150.468 0 140.185 0s-18.84 3.62-25.696 10.848l-21.7 21.416c-7.045 7.043-10.567 15.604-10.567 25.692 0 9.897 3.52 18.56 10.566 25.98L231.544 222.41 92.785 361.168c-7.04 7.043-10.563 15.604-10.563 25.693 0 9.9 3.52 18.566 10.564 25.98l21.7 21.417c7.043 7.043 15.612 10.564 25.697 10.564 10.09 0 18.656-3.52 25.697-10.564L352.025 248.39c7.046-7.423 10.57-16.084 10.57-25.98.002-10.09-3.524-18.655-10.57-25.698z"/>
-</svg>
-
-    </a>
-  </div>
-</article>
-
- 
-
-
-
-
-
-
-
-
-<article class="card tablet:grid-col-4 margin-bottom-4">
-  <div class="card-link" data-href="/what-we-deliver/calc/" tabindex="-1">
-    <div role="img"
-         title="Blue bar graph of various wage estimates"
-         class="card-image-bg "
-         
-         style="background-image: url(/assets/img/projects/calc.png);">
-      
-    </div>
-    <div class="card-banner">
-      
-        <p class="card-tag">General Services Administration</p>
-      
-      
-        <a href="/what-we-deliver/calc/" tabindex="-1" class="card-link-tagline"><h3 class="h4">Better tools for better research</h3></a>
-      
-      <p class="card-description">An efficient way for contracting officers to search past government contracts for fair hourly rates.</p>
-    </div>
-    <a href="/what-we-deliver/calc/" class="card-cta link-arrow-right" tabindex="0">
-      Read more
-      <span class="usa-sr-only">about Better tools for better research</span>
-      <svg class="icon" xmlns="http://www.w3.org/2000/svg" width="444.819" height="444.819" viewBox="0 0 444.819 444.819" role="img" aria-hidden="true">
-  <title>Arrow right</title>
-  <desc>Arrow pointing to the right</desc>
-  <path fill="#046b99" d="M352.025 196.712L165.885 10.848C159.028 3.615 150.468 0 140.185 0s-18.84 3.62-25.696 10.848l-21.7 21.416c-7.045 7.043-10.567 15.604-10.567 25.692 0 9.897 3.52 18.56 10.566 25.98L231.544 222.41 92.785 361.168c-7.04 7.043-10.563 15.604-10.563 25.693 0 9.9 3.52 18.566 10.564 25.98l21.7 21.417c7.043 7.043 15.612 10.564 25.697 10.564 10.09 0 18.656-3.52 25.697-10.564L352.025 248.39c7.046-7.423 10.57-16.084 10.57-25.98.002-10.09-3.524-18.655-10.57-25.698z"/>
-</svg>
-
-    </a>
-  </div>
-</article>
-
-
-    </div>
-  </section>
-</section>
-
-<section class="grid-container usa-section posts_feature">
-  <h2><a href="/blog/">Recent blog posts</a></h2>
-  <ul class="grid-row grid-gap">
-         
-
-<li class="post tablet:grid-col-4">
-  <article>
-    
-    <h3 class="h4 posts_feature-heading">
-      <a class="post-link" href="/2020/09/09/a-federal-guide-to-de-risk-government-technology-projects/"
-        >A federal guide to de-risk government technology projects</a
-      >
-    </h3>
-    <p class="post-excerpt usa-prose">Announcing the federal field guide to de-risk government technology</p>
-    <p>
-      <a
-        class="link-arrow-right post-link-continue_reading"
-        href="/2020/09/09/a-federal-guide-to-de-risk-government-technology-projects/"
-      >
-        Read more
-        <span class="usa-sr-only">about A federal guide to de-risk government technology projects</span>
-        <svg class="icon" xmlns="http://www.w3.org/2000/svg" width="444.819" height="444.819" viewBox="0 0 444.819 444.819" role="img" aria-hidden="true">
-  <title>Arrow right</title>
-  <desc>Arrow pointing to the right</desc>
-  <path fill="#046b99" d="M352.025 196.712L165.885 10.848C159.028 3.615 150.468 0 140.185 0s-18.84 3.62-25.696 10.848l-21.7 21.416c-7.045 7.043-10.567 15.604-10.567 25.692 0 9.897 3.52 18.56 10.566 25.98L231.544 222.41 92.785 361.168c-7.04 7.043-10.563 15.604-10.563 25.693 0 9.9 3.52 18.566 10.564 25.98l21.7 21.417c7.043 7.043 15.612 10.564 25.697 10.564 10.09 0 18.656-3.52 25.697-10.564L352.025 248.39c7.046-7.423 10.57-16.084 10.57-25.98.002-10.09-3.524-18.655-10.57-25.698z"/>
-</svg>
-
-      </a>
-    </p>
-  </article>
-</li>
-      
-
-<li class="post tablet:grid-col-4">
-  <article>
-    
-    <h3 class="h4 posts_feature-heading">
-      <a class="post-link" href="/2020/08/13/10-weeks-improving-government-through-technology/"
-        >10 weeks improving government through technology</a
-      >
-    </h3>
-    <p class="post-excerpt usa-prose">This summer TTS welcomed Abdul Tahlil as a Coding it Forward Civic Digital fellow. Abdul joined the TTS Outreach team to work on the 18F site. We’ve invited Abdul to tell us a bit more about the fellowship program and his experience as an engineer on the team.</p>
-    <p>
-      <a
-        class="link-arrow-right post-link-continue_reading"
-        href="/2020/08/13/10-weeks-improving-government-through-technology/"
-      >
-        Read more
-        <span class="usa-sr-only">about 10 weeks improving government through technology</span>
-        <svg class="icon" xmlns="http://www.w3.org/2000/svg" width="444.819" height="444.819" viewBox="0 0 444.819 444.819" role="img" aria-hidden="true">
-  <title>Arrow right</title>
-  <desc>Arrow pointing to the right</desc>
-  <path fill="#046b99" d="M352.025 196.712L165.885 10.848C159.028 3.615 150.468 0 140.185 0s-18.84 3.62-25.696 10.848l-21.7 21.416c-7.045 7.043-10.567 15.604-10.567 25.692 0 9.897 3.52 18.56 10.566 25.98L231.544 222.41 92.785 361.168c-7.04 7.043-10.563 15.604-10.563 25.693 0 9.9 3.52 18.566 10.564 25.98l21.7 21.417c7.043 7.043 15.612 10.564 25.697 10.564 10.09 0 18.656-3.52 25.697-10.564L352.025 248.39c7.046-7.423 10.57-16.084 10.57-25.98.002-10.09-3.524-18.655-10.57-25.698z"/>
-</svg>
-
-      </a>
-    </p>
-  </article>
-</li>
-      
-
-<li class="post tablet:grid-col-4">
-  <article>
-    
-    <h3 class="h4 posts_feature-heading">
-      <a class="post-link" href="/2020/08/06/doing-user-research-to-design-the-next-gen-wdfn/"
-        >Doing user research to design the next-gen WDFN</a
-      >
-    </h3>
-    <p class="post-excerpt usa-prose">The USGS Water Resources Mission Area is striving to make their water data easier to access for their users. This involves stepping back and reviewing how users currently use waterdata.usgs.gov so that we can make informed decisions moving forward. </p>
-    <p>
-      <a
-        class="link-arrow-right post-link-continue_reading"
-        href="/2020/08/06/doing-user-research-to-design-the-next-gen-wdfn/"
-      >
-        Read more
-        <span class="usa-sr-only">about Doing user research to design the next-gen WDFN</span>
-        <svg class="icon" xmlns="http://www.w3.org/2000/svg" width="444.819" height="444.819" viewBox="0 0 444.819 444.819" role="img" aria-hidden="true">
-  <title>Arrow right</title>
-  <desc>Arrow pointing to the right</desc>
-  <path fill="#046b99" d="M352.025 196.712L165.885 10.848C159.028 3.615 150.468 0 140.185 0s-18.84 3.62-25.696 10.848l-21.7 21.416c-7.045 7.043-10.567 15.604-10.567 25.692 0 9.897 3.52 18.56 10.566 25.98L231.544 222.41 92.785 361.168c-7.04 7.043-10.563 15.604-10.563 25.693 0 9.9 3.52 18.566 10.564 25.98l21.7 21.417c7.043 7.043 15.612 10.564 25.697 10.564 10.09 0 18.656-3.52 25.697-10.564L352.025 248.39c7.046-7.423 10.57-16.084 10.57-25.98.002-10.09-3.524-18.655-10.57-25.698z"/>
-</svg>
-
-      </a>
-    </p>
-  </article>
-</li>
- 
-  </ul>
-</section>
-
-    </div>
-
-    
-      
-
-
-
-
-
-
-
-
-
-        
-
-
-
-
-
-
-
-  
-  <footer class="usa-footer site-footer" role="contentinfo">
-  <div class="footer-section-bottom bg-base-lightest">
-    <div class="grid-container">
-      <div class="grid-row grid-gap padding-y-4">
-        <div class="tablet:grid-col-3">
-          <h4 class="usa-footer__primary-link">Work with us</h4>
-          <ul class="add-list-reset">
-            <li>
-              <a class="usa-footer__secondary-link" href="/about/">About 18F</a>
-            </li>
-            <li>
-              <a class="usa-footer__secondary-link" href="/what-we-deliver">Case studies</a>
-            </li>
-            <li>
-              <a class="usa-footer__secondary-link" href="/how-we-work/">How we work</a>
-            </li>
-            <li>
-              <a class="usa-footer__secondary-link" href="/partnership-principles/">Partnership principles</a>
-            </li>
-          </ul>
-        </div>
-        <div class="tablet:grid-col-3">
-          <h4 class="usa-footer__primary-link">Resources</h4>
-          <ul class="add-list-reset">
-            <li>
-              <a class="usa-footer__secondary-link" href="/blog/">Blog</a>
-            </li>
-            <li>
-                <a class="usa-footer__secondary-link" href="/guides/">Guides</a>
-            </li>
-            <li>
-                <a class="usa-footer__secondary-link" href="https://methods.18f.gov/">18F Methods</a>
-            </li>
-          </ul>
-        </div>
-        <div class="tablet:grid-col-3">
-          <h4 class="usa-footer__primary-link">Policies</h4>
-          <ul class="add-list-reset">
-            <li>
-                <a class="usa-footer__secondary-link" href="/linking-policy/">Linking policy</a>
-            </li>
-            <li>
-                <a class="usa-footer__secondary-link" href="/open-source-policy/">Open source policy</a>
-            </li>
-            <li>
-                <a class="usa-footer__secondary-link" href="/vulnerability-disclosure-policy/">Vulnerability disclosure</a>
-            </li>
-            <li>
-                <a class="usa-footer__secondary-link" href="/code-of-conduct/">Code of conduct</a>
-            </li>
-          </ul>
-        </div>
-        <div class="tablet:grid-col-3">
-          <h4 class="usa-footer__primary-link">Contact us</h4>
-          <ul class="add-list-reset">
-            <li>
-                <a class="usa-footer__secondary-link" href="/contact/">Get in touch</a>
-            </li>
-            <li>
-                <a class="usa-footer__secondary-link" href="/about/#for-press">Press</a>
-            </li>
-            <li>
-                <a class="usa-footer__secondary-link" href="https://github.com/18F/18f.gsa.gov/issues/new">Report a bug</a>
-            </li>
-            <li>
-                <a class="usa-footer__secondary-link" href="/join/">Join 18F</a>
-            </li>
-          </ul>
-        </div>
-      </div>
-    </div>
-  </div>
-  <div class="usa-footer__secondary-section">
-    <div class="grid-container">
-      <div class="grid-row grid-gap">
-        <div class="usa-footer__contact-links mobile-lg:grid-col-12">
-          <div class="usa-footer grid-row grid-gap-1">
-            <div class="grid-col-auto">
-              <a class="usa-social-link usa-social-link--twitter" href="https://twitter.com/18F">
-                <span>Twitter</span>
-              </a>
-            </div>
-            <div class="grid-col-auto">
-              <a class="usa-social-link usa-social-link--github" href="https://github.com/18F">
-                <span>Github</span>
-              </a>
-            </div>
-            <div class="grid-col-auto">
-              <a class="usa-social-link usa-social-link--linkedin" href="https://www.linkedin.com/company/gsa18f/">
-                <span>LinkedIn</span>
-              </a>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</footer>
-
-
-
-<div class="usa-identifier">
-  <section
-    class="usa-identifier__section usa-identifier__section--masthead"
-    aria-label="Agency identifier"
-  >
-    <div class="usa-identifier__container">
-      <div class="usa-identifier__logos">
-        <a href="javascript:void(0);" class="usa-identifier__logo">
-          <img
-            class="usa-identifier__logo-img"
-            src="/assets/img/logos/gsa-logo-w100.png"
-            alt="GSA logo"
-            role="Logo image"
-          />
-        </a>
-      </div>
-      <div class="usa-identifier__identity text-base-lightest" aria-label="Agency description">
-        <p class="usa-identifier__identity-domain">18f.gsa.gov</p>
-        <p class="usa-identifier__identity-disclaimer text-base-lightest">An official website of the
-          <a href="https://www.gsa.gov/about-us/organization/federal-acquisition-service/technology-transformation-services">
-            GSA’s Technology Transformation Services
-          </a>
-        </p>
-      </div>
-    </div>
-  </section>
-  <nav
-    class="usa-identifier__section usa-identifier__section--required-links"
-    aria-label="Important links"
-  >
-    <div class="usa-identifier__container">
-      <ul class="usa-identifier__required-links-list">
-        <li class="usa-identifier__required-links-item">
-          <a
-            class="usa-identifier__required-link usa-link"
-            href="https://www.gsaig.gov/"
-            title="Report fraud, waste, or abuse to the Office of the Inspector General"
-          >
-            Report fraud, waste, or abuse to the Office of the Inspector General
-          </a>
-        </li>
-        <li class="usa-identifier__required-links-item">
-          <a
-            class="usa-identifier__required-link usa-link"
-            href="https://www.gsa.gov/reference/freedom-of-information-act-foia"
-            title="Submit a Freedom of Information Act (FOIA) request"
-          >
-              Submit a Freedom of Information Act (FOIA) request
-          </a>
-        </li>
-        <li class="usa-identifier__required-links-item">
-          <a
-            class="usa-identifier__required-link usa-link"
-            href="https://www.gsa.gov/reference/reports/budget-performance"
-            title="View budget and performance reports"
-          >
-              View budget and performance reports
-          </a>
-        </li>
-        <li class="usa-identifier__required-links-item">
-          <a
-            class="usa-identifier__required-link usa-link"
-            href="https://www.gsa.gov/website-information/accessibility-aids"
-            title="View accessibility statement"
-          >
-            View accessibility statement
-          </a>
-        </li>
-        <li class="usa-identifier__required-links-item">
-          <a
-            class="usa-identifier__required-link usa-link"
-            href="https://www.gsa.gov/about-us/organization/office-of-civil-rights/notification-and-federal-employee-antidiscrimination-and-retaliation-act-of-2002"
-            title="View No FEAR Act data"
-          >
-            View No FEAR Act data
-          </a>
-        </li>
-      </ul>
-    </div>
-  </nav>
-  <section class="usa-identifier__section usa-identifier__section--usagov" aria-label="U.S. government information and services">
-    <div class="usa-identifier__container">
-      <div class="usa-identifier__usagov-description text-base-lightest">Looking for U.S. government information and services?</div>
-      <a href="https://www.usa.gov/" class="usa-link">Visit USA.gov</a>
-    </div>
-  </section>
-</div>
-
-
-
-
-    
-
-    <script src="/assets/js/jquery.min.js"></script>
-<script src="/assets/js/jquery.ajaxchimp.min.js"></script>
-<script src="/assets/js/waypoints.min.js" type="text/javascript"></script>
-<script src="/assets/js/index.js"></script>
-<script src="/assets/js/lib/uswds.min.js"></script>
-<script src="/assets/js/waypoints.js"></script>
-<script src="/assets/js/jekyll-pages-api-search.js"></script>
-<script src="/assets/js/sticky.js"></script>
-
-<!--
-  analytics for DAP, for more info visit
-  https://www.digitalgov.gov/services/dap/
--->
-<script src="https://dap.digitalgov.gov/Universal-Federated-Analytics-Min.js?agency=GSA&subagency=TTS,18F"
-id="_fed_an_ua_tag"></script>
-
-<!--
-  analytics for 18F guild, for more info visited
-  https://github.com/18f/analytics-standards
--->
-<script>
-  (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
-(i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
-m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
-})(window,document,'script','https://www.google-analytics.com/analytics.js','ga');
-
-  ga('create', 'UA-48605964-1', 'auto');
-  ga('set', 'anonymizeIp', true);
-  ga('set', 'forceSSL', true);
-  ga('send', 'pageview');
-</script>
-
-    <script>JEKYLL_PAGES_API_SEARCH_BASEURL = '';</script>
-<script async src="/assets/js/search-bundle.js"></script>
-
-  </body>
-</html>
-`;
-
-const testRobotsTxt = `
-User-agent: *
-Disallow:
-Crawl-delay: 10
-Sitemap: https://18f.gsa.gov/sitemap1.xml
-Sitemap: https://18f.gsa.gov/sitemap2.xml
-`;
-
-const testSitemapXml = `
-
-<?xml version="1.0" encoding="UTF-8"?>
-<urlset xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.sitemaps.org/schemas/sitemap/0.9 http://www.sitemaps.org/schemas/sitemap/0.9/sitemap.xsd" xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
-  <url>
-    <loc>https://18f.gsa.gov/2020/09/09/a-federal-guide-to-de-risk-government-technology-projects/</loc>
-    
-    <lastmod>2020-09-09T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/2020/08/13/10-weeks-improving-government-through-technology/</loc>
-    
-    <lastmod>2020-08-13T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/08/06/doing-user-research-to-design-the-next-gen-wdfn/</loc>
-    
-    <lastmod>2020-08-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/08/04/introducing-the-ux-guide/</loc>
-    
-    <lastmod>2020-08-04T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/07/07/transforming-how-dojs-civil-rights-division-engages-with-the-public/</loc>
-    
-    <lastmod>2020-07-07T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/06/23/building-trust-in-a-public-health-crisis/</loc>
-    
-    <lastmod>2020-06-23T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/06/04/making-critical-government-information-more-resilient/</loc>
-    
-    <lastmod>2020-06-04T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/06/02/design-research-in-critical-times/</loc>
-    
-    <lastmod>2020-06-02T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/05/12/rapid-implementation-of-policy-as-code/</loc>
-    
-    <lastmod>2020-05-12T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/04/23/saving-time-and-improving-data-quality-for-the-national-school-lunch-breakfast-program/</loc>
-    
-    <lastmod>2020-04-23T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/04/21/a-token-of-our-affection-uswds-2/</loc>
-    
-    <lastmod>2020-04-21T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/04/09/18f-help-during-critical-times/</loc>
-    
-    <lastmod>2020-04-09T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/04/01/building-a-collaborative-culture/</loc>
-    
-    <lastmod>2020-04-01T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/03/24/continuous-learning-for-product-managers/</loc>
-    
-    <lastmod>2020-03-24T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/03/19/sixth-anniversary/</loc>
-    
-    <lastmod>2020-03-19T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/03/18/an-acquisition-retrospective/</loc>
-    
-    <lastmod>2020-03-18T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/03/10/ask-18f-po-vs-cor/</loc>
-    
-    <lastmod>2020-03-10T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/03/05/readmes-for-you-and-me/</loc>
-    
-    <lastmod>2020-03-05T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/02/27/areas-for-growth-2-mixed-methods/</loc>
-    
-    <lastmod>2020-02-27T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/02/20/areas-for-growth-part-1-cad/</loc>
-    
-    <lastmod>2020-02-20T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/02/18/forms-resource-federalist/</loc>
-    
-    <lastmod>2020-02-18T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2020/02/06/even-with-a-design-system-you-still-need-a-designer/</loc>
-    
-    <lastmod>2020-02-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/12/10/announcing-the-18f-national-security-and-intelligence-portfolio/</loc>
-    
-    <lastmod>2019-12-10T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/12/03/long-term-teams/</loc>
-    
-    <lastmod>2019-12-03T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/11/21/interior-becoming-a-product-manager-post/</loc>
-    
-    <lastmod>2019-11-21T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/11/19/building-product-management-capacity-in-government-part-2/</loc>
-    
-    <lastmod>2019-11-19T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/11/07/barriers-to-governments-adoption-of-user-centered-design-and-how-to-address-them/</loc>
-    
-    <lastmod>2019-11-07T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/10/29/culture-climate-survey/</loc>
-    
-    <lastmod>2019-10-29T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/08/22/building-product-management-capacity-in-government-part-1/</loc>
-    
-    <lastmod>2019-08-22T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/08/20/an-agile-software-development-solicitation-guide/</loc>
-    
-    <lastmod>2019-08-20T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/08/15/keeping-your-accounts-secure/</loc>
-    
-    <lastmod>2019-08-15T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/08/05/budgeting-handbook/</loc>
-    
-    <lastmod>2019-08-05T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/07/16/on-the-road-to-innovation/</loc>
-    
-    <lastmod>2019-07-16T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/06/20/how-federalist-and-uswds-helps-agencies-become-compliant/</loc>
-    
-    <lastmod>2019-06-20T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/06/18/cross-functional-teams/</loc>
-    
-    <lastmod>2019-06-18T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/05/29/you-might-not-be-as-agile-as-you-think-you-are/</loc>
-    
-    <lastmod>2019-05-29T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/05/21/centralization-gone-right-a-case-study-on-uswds/</loc>
-    
-    <lastmod>2019-05-21T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/05/16/how-18f-and-pif-work-together-in-agencies/</loc>
-    
-    <lastmod>2019-05-16T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/05/09/human-centered-design-for-it-centralization-part-4-what-happens-after-you-centralize/</loc>
-    
-    <lastmod>2019-05-09T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/05/07/human-centered-design-for-it-centralization-part-3-working-with-vendors/</loc>
-    
-    <lastmod>2019-05-07T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/04/11/hcd-for-centralization-series-2-deciding-whether-or-not-to-centralize/</loc>
-    
-    <lastmod>2019-04-11T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/04/09/why-we-love-modular-contracting/</loc>
-    
-    <lastmod>2019-04-09T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/04/08/uswds2-launch-post/</loc>
-    
-    <lastmod>2019-04-08T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/04/03/hcd-for-centralization-series-1-why-prioritizing-users-is-important/</loc>
-    
-    <lastmod>2019-04-03T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/03/26/when-to-use-COTS/</loc>
-    
-    <lastmod>2019-03-26T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/03/22/18F-5-anniv-fav-project/</loc>
-    
-    <lastmod>2019-03-22T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/03/19/18F-5-Anniversary-achieve/</loc>
-    
-    <lastmod>2019-03-19T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/03/07/six-ways-we-improved-tts-design-research-guild/</loc>
-    
-    <lastmod>2019-03-07T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/03/05/the-us-data-federation/</loc>
-    
-    <lastmod>2019-03-05T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/02/28/prerequisites-for-modular-contracting/</loc>
-    
-    <lastmod>2019-02-28T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/02/14/product-manager-10x/</loc>
-    
-    <lastmod>2019-02-14T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/02/12/open-forest-launch-post/</loc>
-    
-    <lastmod>2019-02-12T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/02/07/the-cloud-is-not-a-virtue/</loc>
-    
-    <lastmod>2019-02-07T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2019/02/06/ask-18F-6-analytics-open-source/</loc>
-    
-    <lastmod>2019-02-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/12/20/product-management-at-18F-part-3/</loc>
-    
-    <lastmod>2018-12-20T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/12/11/product-management-at-18F-part-2-acquisitions/</loc>
-    
-    <lastmod>2018-12-11T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/12/04/product-management-at-18f-part-1-new-engagements/</loc>
-    
-    <lastmod>2018-12-04T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/11/20/introduction-to-remote-moderated-usability-testing-part-2-how/</loc>
-    
-    <lastmod>2018-11-20T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/11/14/introduction-to-remote-moderated-usability-testing-part-1/</loc>
-    
-    <lastmod>2018-11-14T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/10/25/modular-contracting-and-working-in-the-open/</loc>
-    
-    <lastmod>2018-10-25T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/10/23/two-exercises-for-improving-design-research-through-reflective-practice/</loc>
-    
-    <lastmod>2018-10-23T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/10/16/exploring-a-new-way-to-make-eligibility-rules-easier-to-implement/</loc>
-    
-    <lastmod>2018-10-16T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/10/09/implementing-rules-without-rules-engines/</loc>
-    
-    <lastmod>2018-10-09T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/09/27/antipatterns-in-agile-contracting/</loc>
-    
-    <lastmod>2018-09-27T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/08/15/shared-infrastructure-as-code/</loc>
-    
-    <lastmod>2018-08-15T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/08/09/etl-extract-transform-learn/</loc>
-    
-    <lastmod>2018-08-09T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/08/07/how-login-used-evidence-based-buying/</loc>
-    
-    <lastmod>2018-08-07T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/08/02/ask-18f-talent/</loc>
-    
-    <lastmod>2018-08-02T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/07/26/what-we-learned-from-building-a-pool-of-agile-vendors/</loc>
-    
-    <lastmod>2018-07-26T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/07/19/taking-the-ato-process-from-6-months-to-30-days/</loc>
-    
-    <lastmod>2018-07-19T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/07/17/five-contracting-tweaks-that-have-yielded-18f-better-procurements/</loc>
-    
-    <lastmod>2018-07-17T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/07/12/the-case-for-open-source-software/</loc>
-    
-    <lastmod>2018-07-12T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/07/10/introducing-accessibility-for-teams/</loc>
-    
-    <lastmod>2018-07-10T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/07/06/ask-18f-plain-language-government-terms/</loc>
-    
-    <lastmod>2018-07-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/07/03/presidential-innovation-fellows-bringing-new-approaches-to-nations-biggest-challenges/</loc>
-    
-    <lastmod>2018-07-03T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/06/14/ask-18f-how-to-shift-the-cultural-mindset-of-my-organization/</loc>
-    
-    <lastmod>2018-06-14T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/06/06/pif-extend-deadline-fall-2018/</loc>
-    
-    <lastmod>2018-06-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/05/24/what-agencies-have-to-say-about-working-in-the-open/</loc>
-    
-    <lastmod>2018-05-24T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/05/22/using-agile-methods-to-improve-the-rfp-process/</loc>
-    
-    <lastmod>2018-05-22T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/05/01/lessons-from-an-18f-product-transition/</loc>
-    
-    <lastmod>2018-05-01T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/04/17/so-youre-a-product-owner/</loc>
-    
-    <lastmod>2018-04-17T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/04/05/ask-18f-form-design/</loc>
-    
-    <lastmod>2018-04-05T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/03/27/four-benefits-to-using-the-full-TTS-tech-stack/</loc>
-    
-    <lastmod>2018-03-27T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/03/13/win-big-by-going-small/</loc>
-    
-    <lastmod>2018-03-13T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/02/28/ask-18f-data-collection/</loc>
-    
-    <lastmod>2018-02-28T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/02/22/new-home-for-the-federal-plain-language-community/</loc>
-    
-    <lastmod>2018-02-22T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/02/21/technology-transformation-services-launches-new-job-site/</loc>
-    
-    <lastmod>2018-02-21T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/02/13/why-your-agency-doesnt-need-a-mobile-app/</loc>
-    
-    <lastmod>2018-02-13T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/02/06/getting-partners-on-board-with-research-findings/</loc>
-    
-    <lastmod>2018-02-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/02/01/moving-a-legacy-microsoft-environment-to-the-azure-cloud/</loc>
-    
-    <lastmod>2018-02-01T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/01/30/getting-prepared-to-prototype/</loc>
-    
-    <lastmod>2018-01-30T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2018/01/25/getting-devops-buy-in/</loc>
-    
-    <lastmod>2018-01-25T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/12/14/getting-stakeholder-buy-in-for-agile-development/</loc>
-    
-    <lastmod>2017-12-14T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/12/12/renata-maziarz-model-civil-servant/</loc>
-    
-    <lastmod>2017-12-12T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/11/30/improving-government-outcomes-through-an-agile-contract-format/</loc>
-    
-    <lastmod>2017-11-30T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/11/08/four-lessons-we-learned-while-building-our-own-design-research-recruiting-tool/</loc>
-    
-    <lastmod>2017-11-08T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/10/23/how-reusable-is-open-source-software/</loc>
-    
-    <lastmod>2017-10-23T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/10/11/pulling-back-the-curtain-on-it-procurement/</loc>
-    
-    <lastmod>2017-10-11T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/10/05/18f-and-federal-laboratories-work-together-to-bring-better-data-to-businesses/</loc>
-    
-    <lastmod>2017-10-05T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/10/03/building-a-large-scale-design-system/</loc>
-    
-    <lastmod>2017-10-03T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/09/27/three-ways-to-manage-research-projects/</loc>
-    
-    <lastmod>2017-09-27T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/09/26/automated-scanning-for-sensitive-information/</loc>
-    
-    <lastmod>2017-09-26T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/09/21/method-cards-tune-up/</loc>
-    
-    <lastmod>2017-09-21T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/09/20/managing-custom-software-development-in-government-when-youre-not-a-software-engineer/</loc>
-    
-    <lastmod>2017-09-20T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/09/18/a-day-in-the-life-of-an-18f-product-owner/</loc>
-    
-    <lastmod>2017-09-18T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/09/12/how-alaska-is-using-transparency/</loc>
-    
-    <lastmod>2017-09-12T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/09/08/forest-service-six-month-check/</loc>
-    
-    <lastmod>2017-09-08T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/09/07/opening-the-nations-crime-data/</loc>
-    
-    <lastmod>2017-09-07T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/09/05/ficam-partners-with-federalist/</loc>
-    
-    <lastmod>2017-09-05T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/08/30/what-makes-a-great-vendor-team/</loc>
-    
-    <lastmod>2017-08-30T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/08/22/government-launches-login-gov/</loc>
-    
-    <lastmod>2017-08-22T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/06/22/why-discovery-research-matters-for-modular-procurement/</loc>
-    
-    <lastmod>2017-06-22T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/06/06/fractal-and-federalist-join-the-us-web-design-standards/</loc>
-    
-    <lastmod>2017-06-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/06/01/federalist-is-out-of-beta-and-open-for-business/</loc>
-    
-    <lastmod>2017-06-01T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/05/30/the-new-fec/</loc>
-    
-    <lastmod>2017-05-30T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/05/25/from-launch-to-landing-how-nasa-took-control-of-its-https-mission/</loc>
-    
-    <lastmod>2017-05-25T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/05/11/the-next-steps-towards-bug-bounty-program-for-technology-transformation-service/</loc>
-    
-    <lastmod>2017-05-11T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/04/27/remembering-larry-matthias-early-partner-excellent-civil-servant/</loc>
-    
-    <lastmod>2017-04-27T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/03/30/how-us-customs-border-protection-uses-us-web-design-standards/</loc>
-    
-    <lastmod>2017-03-30T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/03/21/nasa-journey-with-us-web-design-standards/</loc>
-    
-    <lastmod>2017-03-21T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/03/15/one-agencys-investments-open-source-mean-others-benefit/</loc>
-    
-    <lastmod>2017-03-15T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/03/13/us-web-design-standards-releases-version-10/</loc>
-    
-    <lastmod>2017-03-13T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/02/27/to-get-things-done-you-need-great-secure-tools/</loc>
-    
-    <lastmod>2017-02-27T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/02/02/cloud-gov-is-now-fedramp-authorized/</loc>
-    
-    <lastmod>2017-02-02T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/01/19/the-dark-standup/</loc>
-    
-    <lastmod>2017-01-19T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/01/17/life-changing-magic-writing-release-notes/</loc>
-    
-    <lastmod>2017-01-17T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/01/11/the-best-way-to-build-big-is-to-start-small/</loc>
-    
-    <lastmod>2017-01-11T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/01/10/mark-hopson-negotiating-on-americas-behalf/</loc>
-    
-    <lastmod>2017-01-10T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/01/06/open-source-collaboration-across-agencies-to-improve-https-deployment/</loc>
-    
-    <lastmod>2017-01-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/01/05/building-effective-agile-partnership-between-government-industry/</loc>
-    
-    <lastmod>2017-01-05T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2017/01/04/tracking-the-us-governments-progress-on-moving-https/</loc>
-    
-    <lastmod>2017-01-04T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/12/22/charting-the-future-of-the-draft-us-web-design-standards/</loc>
-    
-    <lastmod>2016-12-22T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/12/15/tts-projects-forward-open-government-partnership-principles/</loc>
-    
-    <lastmod>2016-12-15T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/12/14/how-to-run-an-efficient-meeting/</loc>
-    
-    <lastmod>2016-12-14T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/12/13/lowering-risk-of-experimentation-open-source-code/</loc>
-    
-    <lastmod>2016-12-13T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/12/12/connecting-culture-change-at-18F-and-PBS/</loc>
-    
-    <lastmod>2016-12-12T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/12/02/presidential-innovation-fellows-now-accepting-applications-spring-2017/</loc>
-    
-    <lastmod>2016-12-02T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/11/30/curate-best-thinking-available/</loc>
-    
-    <lastmod>2016-11-30T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/11/25/what-18f-is-thankful-for-this-thanksgiving/</loc>
-    
-    <lastmod>2016-11-25T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/11/22/a-vulnerability-disclosure-policy-for-the-technology-transformation-service/</loc>
-    
-    <lastmod>2016-11-22T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/11/17/leadership-innovation-california-child-welfare-services/</loc>
-    
-    <lastmod>2016-11-17T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/11/15/modular-procurement-state-local-government/</loc>
-    
-    <lastmod>2016-11-15T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/11/07/code-gov-the-next-milestone-federal-open-source-code/</loc>
-    
-    <lastmod>2016-11-07T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/10/28/california-takes-new-approach-procurement/</loc>
-    
-    <lastmod>2016-10-28T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/10/26/writing-for-web-easy-writing-for-users-not/</loc>
-    
-    <lastmod>2016-10-26T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/10/25/three-small-steps-you-can-take-to-reboot-agile-in-your-organization/</loc>
-    
-    <lastmod>2016-10-25T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/10/24/embracing-opportunities-to-improve-and-innovate/</loc>
-    
-    <lastmod>2016-10-24T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/10/14/iterative-workplace-design-denver-federal-center/</loc>
-    
-    <lastmod>2016-10-14T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/10/12/announcing-the-digital-acquisition-accelerator-playbook/</loc>
-    
-    <lastmod>2016-10-12T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/10/11/getting-the-most-juice-out-of-the-squeeze/</loc>
-    
-    <lastmod>2016-10-11T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/10/04/what-is-static-source-analysis/</loc>
-    
-    <lastmod>2016-10-04T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/10/03/using-rfps-make-every-dollar-count/</loc>
-    
-    <lastmod>2016-10-03T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/09/27/usa-gov-updates-vote-gov-now-leaner-faster-multilingual/</loc>
-    
-    <lastmod>2016-09-27T14:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/09/27/pro-tips-for-data-friendly-regulations-proposals/</loc>
-    
-    <lastmod>2016-09-27T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/09/20/mississippi-agile-modular-techniques-child-welfare-system/</loc>
-    
-    <lastmod>2016-09-20T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/09/19/vendors-government-strengthen-partnership-technology-industry-day/</loc>
-    
-    <lastmod>2016-09-19T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/09/15/gao-report-shows-satisfied-customers-areas-for-improvement/</loc>
-    
-    <lastmod>2016-09-15T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/09/12/james-scott-automating-engineering-better-government/</loc>
-    
-    <lastmod>2016-09-12T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/09/06/nicole-fenton-a-wordsmith-joins-government/</loc>
-    
-    <lastmod>2016-09-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/08/31/kanban-for-government/</loc>
-    
-    <lastmod>2016-08-31T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/08/30/sign-up-for-the-technology-industry-day/</loc>
-    
-    <lastmod>2016-08-30T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/08/29/data-act-prototype-simplicty-is-key/</loc>
-    
-    <lastmod>2016-08-29T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/08/26/mark-trammell-unlocking-the-true-potential-of-public-service/</loc>
-    
-    <lastmod>2016-08-26T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/08/25/micro-purchase-design-philosophy-do-one-thing-well/</loc>
-    
-    <lastmod>2016-08-25T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/08/16/what-happens-when-the-whole-team-joins-user-interviews/</loc>
-    
-    <lastmod>2016-08-16T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/08/15/holly-allen-helping-engineer-a-better-18F/</loc>
-    
-    <lastmod>2016-08-15T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/08/10/patterns-for-managing-multi-tenant-cloud-environments/</loc>
-    
-    <lastmod>2016-08-10T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/08/09/the-role-of-bot-bidding-in-micro-purchase-auctions/</loc>
-    
-    <lastmod>2016-08-09T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/08/08/facts-about-publishing-open-source-code-in-government/</loc>
-    
-    <lastmod>2016-08-08T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/08/04/buying-better-digital-products-part-3-mapping-user-stories/</loc>
-    
-    <lastmod>2016-08-04T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/08/02/jacob-harris-from-big-data-journalism-to-micro-purchase-platforms/</loc>
-    
-    <lastmod>2016-08-02T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/07/29/live-streamed-demos-for-exponential-transparency-and-information-sharing/</loc>
-    
-    <lastmod>2016-07-29T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/07/28/bret-mogilefsky-finding-big-good-cloud-gov/</loc>
-    
-    <lastmod>2016-07-28T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/07/26/new-pilot-aims-to-streamline-notice-and-comment-process/</loc>
-    
-    <lastmod>2016-07-26T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/07/26/buying-better-digital-products-part-2-setting-the-product-vision-and-strategy/</loc>
-    
-    <lastmod>2016-07-26T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/07/22/building-empathy-with-stakeholder-interviews-part-2-conversation/</loc>
-    
-    <lastmod>2016-07-22T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/07/21/buying-better-digital-products-part-1-proto-personas-and-understanding-the-problem/</loc>
-    
-    <lastmod>2016-07-21T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/07/19/strategies-for-starting-your-own-writing-lab/</loc>
-    
-    <lastmod>2016-07-19T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/07/18/cloud-gov-full-steam-ahead-fedramp-assessment-process/</loc>
-    
-    <lastmod>2016-07-18T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/07/14/elaine-kamlley-a-developer-commited-diversity/</loc>
-    
-    <lastmod>2016-07-14T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/07/13/technology-transformation-services-looking-for-new-commissioner/</loc>
-    
-    <lastmod>2016-07-13T15:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/07/13/3-ways-for-digital-acqusition-teams-to-work-better/</loc>
-    
-    <lastmod>2016-07-13T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/07/11/conversation-about-static-dynamic-websites/</loc>
-    
-    <lastmod>2016-07-11T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/07/07/when-a-micropurchase-doesnt-work-out-we-try-to-learn-from-it/</loc>
-    
-    <lastmod>2016-07-07T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/07/06/illinois-fourth-grade-class-uses-every-kid-in-a-park-to-explore/</loc>
-    
-    <lastmod>2016-07-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/06/30/6-ways-a-writing-lab-will-help-your-organization/</loc>
-    
-    <lastmod>2016-06-30T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/06/24/5-lessons-in-object-oriented-design-from-sandi-metz/</loc>
-    
-    <lastmod>2016-06-24T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/06/21/using-plain-language-to-bridge-the-gap-between-government-and-industry/</loc>
-    
-    <lastmod>2016-06-21T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/06/20/build-empathy-with-stakeholder-interviews-part-1-preparation/</loc>
-    
-    <lastmod>2016-06-20T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/06/17/exciting-additions-to-analytics-usa-gov/</loc>
-    
-    <lastmod>2016-06-17T12:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/06/16/emanuel-feld-talks-about-his-vizualization-of-government-github-organizations/</loc>
-    
-    <lastmod>2016-06-16T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/06/15/two-agencies-participating-in-the-digital-acquisition-accelerator-pilot/</loc>
-    
-    <lastmod>2016-06-15T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/06/14/prototype-early-prototype-often-lesson-from-the-data-act/</loc>
-    
-    <lastmod>2016-06-14T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/06/07/building-better-by-building-together-with-the-federal-election-commission/</loc>
-    
-    <lastmod>2016-06-07T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/06/02/open-data-democratizes-innovation/</loc>
-    
-    <lastmod>2016-06-02T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/05/31/taking-an-agile-approach-to-content/</loc>
-    
-    <lastmod>2016-05-31T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/05/24/the-user-centered-redesign-of-identitytheft-gov/</loc>
-    
-    <lastmod>2016-05-24T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/05/23/a-sestina-on-sunsetting-content/</loc>
-    
-    <lastmod>2016-05-23T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/05/20/get-your-pass-for-kids-to-parks-day/</loc>
-    
-    <lastmod>2016-05-20T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/05/19/content-debt-what-it-is-where-to-find-it-and-how-to-prevent-it-in-the-first-place/</loc>
-    
-    <lastmod>2016-05-19T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/05/18/why-were-moving-18f-gsa-gov-to-federalist/</loc>
-    
-    <lastmod>2016-05-18T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/05/17/reimagining-federal-websites-with-the-draft-us-web-design-standards/</loc>
-    
-    <lastmod>2016-05-17T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/05/13/how-18f-handles-information-security-and-third-party-applications/</loc>
-    
-    <lastmod>2016-05-13T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/05/12/introducing-18fs-new-visual-identity/</loc>
-    
-    <lastmod>2016-05-12T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/05/11/tackling-challenges-facing-cities-in-collaboration-with-all-levels-of-government/</loc>
-    
-    <lastmod>2016-05-11T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/05/10/building-a-modern-shared-authentication-platform/</loc>
-    
-    <lastmod>2016-05-10T14:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/05/10/gsa-invests-in-18fs-work-with-new-service/</loc>
-    
-    <lastmod>2016-05-10T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/05/06/the-day-90-kids-came-to-code-with-18f/</loc>
-    
-    <lastmod>2016-05-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/05/03/delivering-the-next-generation-of-digital-government/</loc>
-    
-    <lastmod>2016-05-03T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/05/02/from-spreadsheet-to-api-to-app-a-better-contract-forecast-tool/</loc>
-    
-    <lastmod>2016-05-02T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/04/28/a-guide-to-the-18f-writing-lab/</loc>
-    
-    <lastmod>2016-04-28T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/04/27/making-a-distributed-design-team-work/</loc>
-    
-    <lastmod>2016-04-27T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/04/26/thinking-about-the-future-of-the-post-office-an-interview-with-amanda-weaver/</loc>
-    
-    <lastmod>2016-04-26T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/04/25/lean-on-me-asking-for-help-on-the-content-team/</loc>
-    
-    <lastmod>2016-04-25T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/04/22/what-is-an-api/</loc>
-    
-    <lastmod>2016-04-22T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/04/21/checklistomania-makes-it-easy-to-keep-track-of-relative-tasks/</loc>
-    
-    <lastmod>2016-04-21T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/04/20/a-clear-audience-makes-for-a-good-blog-post/</loc>
-    
-    <lastmod>2016-04-20T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/04/19/looking-at-the-different-ways-to-test-content/</loc>
-    
-    <lastmod>2016-04-19T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/04/18/the-teams-they-are-a-changin/</loc>
-    
-    <lastmod>2016-04-18T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/04/15/compliance-masonry-buildling-a-risk-management-platform/</loc>
-    
-    <lastmod>2016-04-15T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/04/13/7-18f-projects-that-state-and-local-governments-can-reuse/</loc>
-    
-    <lastmod>2016-04-13T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/04/08/how-we-get-high-availability-with-elasticsearch-and-ruby-on-rails/</loc>
-    
-    <lastmod>2016-04-08T12:05:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/04/08/making-more-consistent-decisions-with-design-principles-a-new-18f-guide/</loc>
-    
-    <lastmod>2016-04-08T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/04/06/take-our-code-18f-projects-you-can-reuse/</loc>
-    
-    <lastmod>2016-04-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/04/05/three-teams-using-the-draft-us-web-design-standards-talk-about-their-experiences/</loc>
-    
-    <lastmod>2016-04-05T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/04/04/national-earthquake-information-centers-open-source-team/</loc>
-    
-    <lastmod>2016-04-04T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/04/01/developing-the-draft-us-web-design-standards-ui-components/</loc>
-    
-    <lastmod>2016-04-01T13:30:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/04/01/what-10-weeks-at-18f-taught-me/</loc>
-    
-    <lastmod>2016-04-01T10:30:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/31/new-guide-provides-tools-for-product-leads/</loc>
-    
-    <lastmod>2016-03-31T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/30/how-we-shared-a-visual-style-across-multiple-sites/</loc>
-    
-    <lastmod>2016-03-30T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/29/best-practices-for-building-an-accessible-website-using-the-draft-us-web-design-standards/</loc>
-    
-    <lastmod>2016-03-29T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/28/interesting-things-we-learned-from-examining-traffic-patterns-on-analytics-usa-gov/</loc>
-    
-    <lastmod>2016-03-28T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/25/how-design-consistency-helps-users-navigate-federal-websites/</loc>
-    
-    <lastmod>2016-03-25T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/24/video-highlights-from-18f-speeches-over-the-last-year/</loc>
-    
-    <lastmod>2016-03-24T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/23/what-have-you-loved-working-on/</loc>
-    
-    <lastmod>2016-03-23T15:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/23/how-to-integrate-the-draft-us-web-design-standards-into-existing-projects/</loc>
-    
-    <lastmod>2016-03-23T13:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/22/what-was-your-path-to-18f/</loc>
-    
-    <lastmod>2016-03-22T18:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/22/helping-california-buy-a-new-child-welfare-system/</loc>
-    
-    <lastmod>2016-03-22T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/21/we-asked-100-of-our-coworkers-why-did-you-join-18f/</loc>
-    
-    <lastmod>2016-03-21T16:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/18/18f-year-two-by-the-numbers/</loc>
-    
-    <lastmod>2016-03-18T13:50:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/18/how-user-archetypes-informed-the-draft-us-web-design-standards/</loc>
-    
-    <lastmod>2016-03-18T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/16/making-the-draft-web-design-standards-better-through-feedback/</loc>
-    
-    <lastmod>2016-03-16T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/15/treasury-and-the-data-act-full-of-sunshine/</loc>
-    
-    <lastmod>2016-03-15T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/11/designing-services-that-are-accessible-transparent-and-easy-to-use/</loc>
-    
-    <lastmod>2016-03-11T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/10/announcing-the-launch-of-the-digital-acquisitions-pilot/</loc>
-    
-    <lastmod>2016-03-10T14:45:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/10/what-people-think-about-before-sharing-personal-information/</loc>
-    
-    <lastmod>2016-03-10T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/09/trust-as-a-two-way-street/</loc>
-    
-    <lastmod>2016-03-09T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/08/san-diego-summit-opens-doors-to-technology-partnerships/</loc>
-    
-    <lastmod>2016-03-08T13:22:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/08/too-many-options-make-complicated-decisions-harder/</loc>
-    
-    <lastmod>2016-03-08T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/07/the-need-for-transparency-in-government/</loc>
-    
-    <lastmod>2016-03-07T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/04/how-people-use-proxies-to-interact-with-the-government/</loc>
-    
-    <lastmod>2016-03-04T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/03/five-questions-for-the-nasa-chief-historian/</loc>
-    
-    <lastmod>2016-03-03T03:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/03/how-people-learn-to-navigate-government-services/</loc>
-    
-    <lastmod>2016-03-03T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/02/strategies-people-use-when-interacting-with-the-federal-government/</loc>
-    
-    <lastmod>2016-03-02T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/03/01/what-we-learned-after-interviewing-people-about-their-interactions-with-the-federal-government/</loc>
-    
-    <lastmod>2016-03-01T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/02/29/informing-the-future-of-the-federal-front-door/</loc>
-    
-    <lastmod>2016-02-29T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/02/24/two-new-open-positions-to-guide-our-business-and-our-people/</loc>
-    
-    <lastmod>2016-02-24T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/02/23/software-maintenance-is-an-anti-pattern/</loc>
-    
-    <lastmod>2016-02-23T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/02/19/how-and-why-we-built-the-micro-purchase-platform/</loc>
-    
-    <lastmod>2016-02-19T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/02/18/analytics.usa.gov-agency-specific-dashboards/</loc>
-    
-    <lastmod>2016-02-18T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/02/12/happy-valentines-day-from-18F/</loc>
-    
-    <lastmod>2016-02-12T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/02/09/bringing-iterative-development-to-our-talent-team/</loc>
-    
-    <lastmod>2016-02-09T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/02/09/tips-for-capturing-the-best-data-from-user-interviews/</loc>
-    
-    <lastmod>2016-02-09T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/02/01/new-us-digital-registry-authenticates-official-public-service-accounts/</loc>
-    
-    <lastmod>2016-02-01T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/01/25/how-user-story-estimation-helps-my-team-deliver-value/</loc>
-    
-    <lastmod>2016-01-25T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/01/22/18f-writing-lab/</loc>
-    
-    <lastmod>2016-01-22T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/01/21/what-we-can-learn-from-the-interiors-social-feeds/</loc>
-    
-    <lastmod>2016-01-21T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/01/15/micro-purchase-auctions-round-2-what-we-learned/</loc>
-    
-    <lastmod>2016-01-15T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/01/11/introducing-the-css-coding-style-guide/</loc>
-    
-    <lastmod>2016-01-11T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/01/07/18f-new-years-resolution-be-even-more-open/</loc>
-    
-    <lastmod>2016-01-07T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/01/07/announcing-the-18f-micro-purchase-platform/</loc>
-    
-    <lastmod>2016-01-07T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/01/06/tips-for-adapting-analytics-usa-gov/</loc>
-    
-    <lastmod>2016-01-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/01/05/turning-learning-up-to-11-knowledge-sharing/</loc>
-    
-    <lastmod>2016-01-05T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2016/01/04/turning-learning-up-to-11-transparency/</loc>
-    
-    <lastmod>2016-01-04T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/12/30/turning-learning-up-to-11/</loc>
-    
-    <lastmod>2015-12-30T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/12/29/is-your-project-using-agilefall/</loc>
-    
-    <lastmod>2015-12-29T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/12/23/looking-back-2015-our-own-words/</loc>
-    
-    <lastmod>2015-12-23T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/12/22/uk-digital-service-visits-us-to-begin-series-of-exchanges/</loc>
-    
-    <lastmod>2015-12-22T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/12/17/the-agile-bpa-is-ready-to-launch/</loc>
-    
-    <lastmod>2015-12-17T14:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/12/17/plain-language-betafec-new-content-design/</loc>
-    
-    <lastmod>2015-12-17T12:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/12/16/useiti-whats-in-first-report/</loc>
-    
-    <lastmod>2015-12-16T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/12/15/how-bot-named-dolores-landingham-transformed-18fs-onboarding/</loc>
-    
-    <lastmod>2015-12-15T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/12/11/how-we-test-18f-gsa-gov/</loc>
-    
-    <lastmod>2015-12-11T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/12/09/an-open-source-government-is-a-faster-more-efficient-government/</loc>
-    
-    <lastmod>2015-12-09T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/12/08/learning-how-to-build-a-better-front-door-for-the-federal-government/</loc>
-    
-    <lastmod>2015-12-08T14:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/12/08/using-emoji-for-knowledge-sharing/</loc>
-    
-    <lastmod>2015-12-08T10:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/12/07/what-exactly-do-we-even-do-all-day/</loc>
-    
-    <lastmod>2015-12-07T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/12/03/power-of-mindset-normatives-matthew-milan/</loc>
-    
-    <lastmod>2015-12-03T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/12/03/epa-environmental-digital-services-marketplace/</loc>
-    
-    <lastmod>2015-12-03T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/12/02/analytics-usa-gov-new-features-more-data/</loc>
-    
-    <lastmod>2015-12-02T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/12/01/how-we-dramatically-improved-18fs-onboarding-process-in-3-months/</loc>
-    
-    <lastmod>2015-12-01T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/11/30/complex-data-in-useiti/</loc>
-    
-    <lastmod>2015-11-30T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/11/20/how-we-use-a-lean-approach-to-product-design/</loc>
-    
-    <lastmod>2015-11-20T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/11/19/delivery-partnership-playbook/</loc>
-    
-    <lastmod>2015-11-19T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/11/18/the-current-future-of-18f-marketplaces/</loc>
-    
-    <lastmod>2015-11-18T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/11/18/automating-easy-government-decisions-with-machine-learning/</loc>
-    
-    <lastmod>2015-11-18T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/11/17/gsa-gov-refreshed-with-eye-toward-mobile-users/</loc>
-    
-    <lastmod>2015-11-17T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/11/17/choose-design-over-architecture/</loc>
-    
-    <lastmod>2015-11-17T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/11/13/answering-common-questions-about-cloud-gov/</loc>
-    
-    <lastmod>2015-11-13T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/11/10/boston-is-using-gsa-calc-tool/</loc>
-    
-    <lastmod>2015-11-10T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/11/06/conversation-with-cfpb-natalie-kurz/</loc>
-    
-    <lastmod>2015-11-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/11/06/micro-purchase-lessons/</loc>
-    
-    <lastmod>2015-11-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/11/04/complexity-is-the-adversary/</loc>
-    
-    <lastmod>2015-11-04T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/11/03/content-strategy-for-all-insights-from-cfpb-natalie-kurz/</loc>
-    
-    <lastmod>2015-11-03T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/11/02/useiti-what-we-learned-where-were-headed/</loc>
-    
-    <lastmod>2015-11-02T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/10/29/welcome-to-betafec/</loc>
-    
-    <lastmod>2015-10-29T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/10/26/micro-purchase-criteria-announcement/</loc>
-    
-    <lastmod>2015-10-26T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/10/22/preventing-technical-debt/</loc>
-    
-    <lastmod>2015-10-22T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/10/21/every-kid-in-a-park-follow-up/</loc>
-    
-    <lastmod>2015-10-21T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/10/20/digital-analytics-program-anniversary/</loc>
-    
-    <lastmod>2015-10-20T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/10/15/best-practices-for-distributed-teams/</loc>
-    
-    <lastmod>2015-10-15T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/10/14/whats-in-a-name-understanding-and-using-gov-acronyms/</loc>
-    
-    <lastmod>2015-10-14T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/10/13/open-source-micropurchasing/</loc>
-    
-    <lastmod>2015-10-13T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/10/09/cloud-gov-launch/</loc>
-    
-    <lastmod>2015-10-09T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/10/08/updated-18f-content-guide/</loc>
-    
-    <lastmod>2015-10-08T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/10/07/digital-economy-practice/</loc>
-    
-    <lastmod>2015-10-07T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/10/06/how-we-start-a-new-project/</loc>
-    
-    <lastmod>2015-10-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/10/05/managing-technical-debt/</loc>
-    
-    <lastmod>2015-10-05T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/10/02/building-a-better-government-design-team/</loc>
-    
-    <lastmod>2015-10-02T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/09/28/web-design-standards/</loc>
-    
-    <lastmod>2015-09-28T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/09/22/usa-gov-launches-vote-usa-gov-to-help-citizens-register-and-connect-with-states/</loc>
-    
-    <lastmod>2015-09-22T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/09/21/cfpb-design-speaker-series/</loc>
-    
-    <lastmod>2015-09-21T16:52:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/09/21/new-citizenship-resources/</loc>
-    
-    <lastmod>2015-09-21T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/09/15/federalist-platform-launch/</loc>
-    
-    <lastmod>2015-09-15T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/09/14/college-scorecard-launch/</loc>
-    
-    <lastmod>2015-09-14T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/09/09/how-a-two-day-sprint-moved-an-agency-twenty-years-forward/</loc>
-    
-    <lastmod>2015-09-09T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/09/08/reimagining-the-immigration-process/</loc>
-    
-    <lastmod>2015-09-08T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/09/04/what-is-technical-debt/</loc>
-    
-    <lastmod>2015-09-04T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/09/03/every-kid-in-a-park/</loc>
-    
-    <lastmod>2015-09-03T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/09/01/govconnect-launch/</loc>
-    
-    <lastmod>2015-09-01T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/08/31/how-playing-with-legos-taught-executives-agile/</loc>
-    
-    <lastmod>2015-08-31T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/08/28/announcing-the-agile-BPA-awards/</loc>
-    
-    <lastmod>2015-08-28T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/08/26/aaron-gustafson-beyond-responsive-design/</loc>
-    
-    <lastmod>2015-08-26T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/08/18/pif-made-permanent/</loc>
-    
-    <lastmod>2015-08-18T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/08/10/18f-design-methods/</loc>
-    
-    <lastmod>2015-08-10T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/08/07/technical-debt-1/</loc>
-    
-    <lastmod>2015-08-07T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/08/06/communicart-tool-will-streamline-purchase-card-process/</loc>
-    
-    <lastmod>2015-08-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/07/29/style-guide-for-open-source-documentation/</loc>
-    
-    <lastmod>2015-07-29T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/07/17/the-new-10/</loc>
-    
-    <lastmod>2015-07-17T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/07/16/introduction-to-https-webinar/</loc>
-    
-    <lastmod>2015-07-16T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/07/15/openfec-api-update/</loc>
-    
-    <lastmod>2015-07-15T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/07/08/openfec-api/</loc>
-    
-    <lastmod>2015-07-08T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/07/06/18f-content-guide/</loc>
-    
-    <lastmod>2015-07-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/06/25/improving-healthcare-gov-experience/</loc>
-    
-    <lastmod>2015-06-25T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/06/24/thomas-vander-wal-event/</loc>
-    
-    <lastmod>2015-06-24T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/06/22/avoiding-cloudfall/</loc>
-    
-    <lastmod>2015-06-22T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/06/17/what-estonia-can-teach-us-about-government-records/</loc>
-    
-    <lastmod>2015-06-17T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/06/16/fast-company-story/</loc>
-    
-    <lastmod>2015-06-16T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/06/15/agile-bpa-is-here/</loc>
-    
-    <lastmod>2015-06-15T12:30:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/06/15/building-a-better-welcome-wagon/</loc>
-    
-    <lastmod>2015-06-15T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/06/11/18f-at-national-civic-hacking-day/</loc>
-    
-    <lastmod>2015-06-11T11:30:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/06/11/moveforward-map-give-video/</loc>
-    
-    <lastmod>2015-06-11T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/06/09/data-act-data-act-explainer/</loc>
-    
-    <lastmod>2015-06-09T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/06/08/the-us-government-is-moving-to-https-everywhere/</loc>
-    
-    <lastmod>2015-06-08T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/06/03/giving-back-to-open-source/</loc>
-    
-    <lastmod>2015-06-03T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/06/02/taking-the-pulse-of-the-federal-governments-web-presence/</loc>
-    
-    <lastmod>2015-06-02T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/05/29/moveforward-analytics-video/</loc>
-    
-    <lastmod>2015-05-29T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/05/28/18F-guides/</loc>
-    
-    <lastmod>2015-05-28T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/05/27/dan-brown-conflict-in-design/</loc>
-    
-    <lastmod>2015-05-27T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/05/21/tockingtime/</loc>
-    
-    <lastmod>2015-05-21T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/05/19/five-questions-with-dan-brown/</loc>
-    
-    <lastmod>2015-05-19T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/05/18/myusa/</loc>
-    
-    <lastmod>2015-05-18T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/05/14/18Fpages/</loc>
-    
-    <lastmod>2015-05-14T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/05/12/on-culture-change-a-code-of-conduct/</loc>
-    
-    <lastmod>2015-05-12T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/05/12/announcing-the-calc-tool/</loc>
-    
-    <lastmod>2015-05-12T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/05/11/moveforward/</loc>
-    
-    <lastmod>2015-05-11T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/05/08/layering-innovation/</loc>
-    
-    <lastmod>2015-05-08T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/05/04/pair-programming-why-two-heads-are-better-than-one/</loc>
-    
-    <lastmod>2015-05-04T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/04/29/18f-how-we-write/</loc>
-    
-    <lastmod>2015-04-29T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/04/28/intersection-of-art-and-technology/</loc>
-    
-    <lastmod>2015-04-28T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/04/24/agile-developments-secret-weapon-transparency/</loc>
-    
-    <lastmod>2015-04-24T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/04/23/the-dat-team-talks-data-streams/</loc>
-    
-    <lastmod>2015-04-23T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/04/23/coming-soon-the-agile-delivery-services-soliciatation/</loc>
-    
-    <lastmod>2015-04-23T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/04/21/hackathons-not-just-for-folks-who-code/</loc>
-    
-    <lastmod>2015-04-21T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/04/20/act-iac-event-on-devops-in-the-government/</loc>
-    
-    <lastmod>2015-04-20T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/04/09/flexibility-when-releasing-a-new-product-peace-corps-new-donation-platform/</loc>
-    
-    <lastmod>2015-04-09T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/04/08/day-in-the-life-of-talent-manager/</loc>
-    
-    <lastmod>2015-04-08T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/04/07/icymi-a11yhack-accessibility-awareness-hackathon/</loc>
-    
-    <lastmod>2015-04-07T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/04/03/how-to-welcome-new-coders-to-a-civic-hackathon/</loc>
-    
-    <lastmod>2015-04-03T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/04/03/18f-discussion-should-project-teams-code-first-or-design-first/</loc>
-    
-    <lastmod>2015-04-03T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/04/02/govtechhack-hacking-for-civic-improvement/</loc>
-    
-    <lastmod>2015-04-02T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/03/31/focus-on-accessibility/</loc>
-    
-    <lastmod>2015-03-31T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/03/30/new-rfp-ghostwriting-service-to-improve-contract-success/</loc>
-    
-    <lastmod>2015-03-30T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/03/24/making-twitter-images-more-accessible/</loc>
-    
-    <lastmod>2015-03-24T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/03/20/one-year-in-and-looking-forward/</loc>
-    
-    <lastmod>2015-03-20T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/03/19/18f-by-the-numbers/</loc>
-    
-    <lastmod>2015-03-19T18:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/03/19/how-we-built-analytics-usa-gov/</loc>
-    
-    <lastmod>2015-03-19T10:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/03/18/sunshine-week-extractive-industries-transparency-initiative-event/</loc>
-    
-    <lastmod>2015-03-18T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/03/17/does-18f-pass-the-bechdel-test-for-tech/</loc>
-    
-    <lastmod>2015-03-17T14:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/03/17/for-public-comment-the-https-only-standard/</loc>
-    
-    <lastmod>2015-03-17T10:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/03/13/how-to-protosketch/</loc>
-    
-    <lastmod>2015-03-13T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/03/12/day-in-the-life-of-an-18F-content-designer/</loc>
-    
-    <lastmod>2015-03-12T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/03/11/onet-today-and-beyond/</loc>
-    
-    <lastmod>2015-03-11T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/03/10/Labor-Category-Descriptions-for-Agile-Procurements/</loc>
-    
-    <lastmod>2015-03-10T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/03/06/ux-lessons-learned-from-a-procurement-project/</loc>
-    
-    <lastmod>2015-03-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/03/05/making-procurement-easier-some-questions-for-developer-kaitlin-devine/</loc>
-    
-    <lastmod>2015-03-05T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/03/04/annoucing-oasis-discovery-making-market-research-easier/</loc>
-    
-    <lastmod>2015-03-04T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/03/03/how-to-use-github-and-the-terminal-a-guide/</loc>
-    
-    <lastmod>2015-03-03T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/03/02/cos-post/</loc>
-    
-    <lastmod>2015-03-02T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/02/26/Five-Ways-To-Turn-A-Bad-Idea-Into-A-Great-One/</loc>
-    
-    <lastmod>2015-02-26T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/02/25/We-Are-Hiring/</loc>
-    
-    <lastmod>2015-02-25T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/02/18/steve-portigal-pre-event-qa-post/</loc>
-    
-    <lastmod>2015-02-18T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/02/17/three-18f-products-that-will-help-your-workplace/</loc>
-    
-    <lastmod>2015-02-17T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/02/12/highlights-from-the-agile-delivery-services-industry-day-events/</loc>
-    
-    <lastmod>2015-02-12T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/02/11/a-story-of-an-agile-workshop/</loc>
-    
-    <lastmod>2015-02-11T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/02/09/the-first-gov-domains-hardcoded-into-your-browser-as-all-https/</loc>
-    
-    <lastmod>2015-02-09T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/01/21/join-us/</loc>
-    
-    <lastmod>2015-01-21T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/01/16/open-source-for-good-government/</loc>
-    
-    <lastmod>2015-01-16T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/01/13/an-open-source-tool-for-easier-database-testing/</loc>
-    
-    <lastmod>2015-01-13T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/01/08/creating-a-federal-marketplace-for-agile-delivery-services/</loc>
-    
-    <lastmod>2015-01-08T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2015/01/06/protosketch/</loc>
-    
-    <lastmod>2015-01-06T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/12/23/hub/</loc>
-    
-    <lastmod>2014-12-23T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/12/18/a-complete-list-of-gov-domains/</loc>
-    
-    <lastmod>2014-12-18T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/12/17/snippets/</loc>
-    
-    <lastmod>2014-12-17T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/12/15/going-to-gov-college/</loc>
-    
-    <lastmod>2014-12-15T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/12/11/large-scale-development-culture-change/</loc>
-    
-    <lastmod>2014-12-11T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/12/10/great-community-turnout-for-open-source-hack-night/</loc>
-    
-    <lastmod>2014-12-10T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/11/26/how-to-use-more-open-source/</loc>
-    
-    <lastmod>2014-11-26T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/11/17/taking-control-of-our-website-with-jekyll-and-webhooks/</loc>
-    
-    <lastmod>2014-11-17T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/11/13/why-we-use-https-in-every-gov-website-we-make/</loc>
-    
-    <lastmod>2014-11-13T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/10/21/how-to-run-your-own-3-sprint-agile-workshop/</loc>
-    
-    <lastmod>2014-10-21T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/10/01/open-source-hack-series-midas/</loc>
-    
-    <lastmod>2014-10-01T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/09/25/design-studio-onrr/</loc>
-    
-    <lastmod>2014-09-25T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/09/18/getting-to-work-for-the-american-people/</loc>
-    
-    <lastmod>2014-09-18T19:21:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/09/15/say-hello-to-the-new-presidential-innovation-fellows/</loc>
-    
-    <lastmod>2014-09-15T18:39:05+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/09/08/the-encasement-strategy-on-legacy-systems-and-the/</loc>
-    
-    <lastmod>2014-09-08T00:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/09/04/a-new-look-at-the-freedom-of-information-act/</loc>
-    
-    <lastmod>2014-09-04T16:02:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/08/21/creating-an-open-fec/</loc>
-    
-    <lastmod>2014-08-21T16:10:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/08/12/the-contributors-guide-to-18f-code-for-the-common/</loc>
-    
-    <lastmod>2014-08-12T16:31:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/07/31/working-in-public-from-day-1/</loc>
-    
-    <lastmod>2014-07-31T17:01:47+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/07/29/18f-an-open-source-team/</loc>
-    
-    <lastmod>2014-07-29T18:01:55+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/07/23/take-a-gander-at-our-developer-page/</loc>
-    
-    <lastmod>2014-07-23T14:01:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/07/16/midas-a-marketplace-for-innovation-in-government/</loc>
-    
-    <lastmod>2014-07-16T15:01:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/07/15/hot-off-the-press-18fs-api-standards/</loc>
-    
-    <lastmod>2014-07-15T18:41:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/06/25/intro-to-apis-working-with-urls-json-apis-and-open/</loc>
-    
-    <lastmod>2014-06-25T15:48:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/05/29/announcing-the-developer-program-a-new-hub-for/</loc>
-    
-    <lastmod>2014-05-29T21:15:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/05/19/packaging-up-api-usability-testing-for-agency-re-use/</loc>
-    
-    <lastmod>2014-05-19T15:02:14+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/05/16/slides-from-the-inaugural-18f-demo-day/</loc>
-    
-    <lastmod>2014-05-16T14:02:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/05/14/hacking-bureaucracy-improving-hiring-and-software/</loc>
-    
-    <lastmod>2014-05-14T15:02:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/05/11/18f-shows-what-is-possible-in-government-with-fbopen/</loc>
-    
-    <lastmod>2014-05-11T17:02:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/05/10/make-gov-apis-better-with-user-experience/</loc>
-    
-    <lastmod>2014-05-10T17:01:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/05/09/a-few-notes-on-notalone-gov/</loc>
-    
-    <lastmod>2014-05-09T17:49:34+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/04/28/18f-demo-day-may-9-2014/</loc>
-    
-    <lastmod>2014-04-28T15:59:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/04/12/how-a-pepperoni-pizza-is-inspiring-open-government/</loc>
-    
-    <lastmod>2014-04-12T16:04:11+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/04/11/open-source-terms-of-service-a-better-developer/</loc>
-    
-    <lastmod>2014-04-11T15:07:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/04/01/ask-us-almost-anything/</loc>
-    
-    <lastmod>2014-04-01T19:44:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/03/31/announcing-fbopen-government-opportunities-made-easier/</loc>
-    
-    <lastmod>2014-03-31T15:02:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/03/27/uncle-sam-wants-you-to-be-a-hero-hacker/</loc>
-    
-    <lastmod>2014-03-27T19:37:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/03/21/29-minutes/</loc>
-    
-    <lastmod>2014-03-21T16:00:00+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/03/19/hello-world-we-are-18f/</loc>
-    
-    <lastmod>2014-03-19T14:14:29+00:00</lastmod>
-    
-  </url>
-  
-
-  <url>
-    <loc>https://18f.gsa.gov/2014/03/12/coming-soon/</loc>
-    
-    <lastmod>2014-03-12T14:11:00+00:00</lastmod>
-    
-  </url>
-  
-  
-  <url>
-    <loc>https://18f.gsa.gov/404/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/500/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/author/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/code-of-conduct/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/contact/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/foia-reading-room/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/guides/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/how-we-work/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/admin/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/linking-policy/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/open-source-policy/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/dashboard/stages/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/thank-you/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/vulnerability-disclosure-policy/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/what-we-deliver/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/18f/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/transformation-services/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/culture-change/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/modern-practices/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/presidential-innovation-fellows/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/talent/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/api/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/procurement/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/fbopen/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/acquisition-services/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/how-we-work/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/events/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/best-practices/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/open-source/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/user-centered-design/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/demo-day/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/culture/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/agency-work/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/usability/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/join-us/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/open-opportunities/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/evangelism/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/education/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/workshop/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/midas/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/platforms/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/fec-gov/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/open-data/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/foia/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/department-of-justice/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/training/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/useiti/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/hackathons/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/agile/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/security/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/https/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/18f-gsa-gov/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/communication-tools-and-practices/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/jekyll/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/open-government/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/collaboration-tools/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/navy-reserve/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/agile-bpa/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/testing/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/tools-you-can-use/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/gov-uk/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/u-s-digital-service/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/speaker-series/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/design/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/user-research/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/discovery-gsa-gov/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/lessons-learned/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/day-in-the-life/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/content-design/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/staff-profiles/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/transparency/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/analytics-usa-gov/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/accessibility/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/rfp-ghostwriting/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/peace-corps/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/devops/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/cloud-gov/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/digital-services-movement/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/general-services-administration/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/calc/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/product-launch/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/myusa/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/guides/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/video/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/pulse-cio-gov/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/hub/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/data-act/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/onboarding/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/treasury/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/new-ten/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/c2/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/technical-debt/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/every-kid-in-a-park/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/uscis/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/department-of-labor/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/college-scorecard/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/federalist/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/vote-usa-gov/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/web-design-system/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/distributed/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/micro-purchase-platforms/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/digitalgov-community/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/machine-learning/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/data-access/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/environmental-protection-agency/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/convincing-stakeholders/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/federal-front-door/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/eregulations/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/interview/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/digital-acquisition-accelerator/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/second-anniversary/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/state-and-local-practice/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/modular-contracting/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/health-and-human-services/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/u-s-geological-survey/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/technical-guides/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/technology-transformation-services/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/login-gov/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/identity/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/public-buildings-service/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/department-of-commerce/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/national-technical-information-service/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/code-gov/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/product/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/nasa/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/forest-service/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/fbi/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/alaska/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/encasement/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/legacy/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/ask-18f/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/gsa/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/10x/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/anniversary/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/it-centralization/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/national-security/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/portfolios/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/rules-as-code/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/form-design/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/case-management-system/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/tags/civil-rights/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/search/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/2/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/3/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/4/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/5/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/6/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/7/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/8/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/9/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/10/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/11/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/12/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/13/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/14/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/15/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/16/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/17/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/18/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/19/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/20/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/21/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/22/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/23/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/24/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/25/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/26/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/27/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/28/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/29/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/30/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/31/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/32/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/33/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/34/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/35/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/36/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/37/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/38/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/39/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/40/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/41/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/42/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/43/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/44/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/45/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/46/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/47/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/48/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/49/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/50/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/51/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/52/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/53/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/54/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/55/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/56/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/57/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/58/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/59/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/60/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/61/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/62/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/63/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/64/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/65/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/66/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/67/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/68/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/69/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/70/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/71/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/72/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/73/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/74/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/75/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/76/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/77/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/78/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/79/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/80/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/81/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/82/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/83/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/84/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/85/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/86/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/87/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/88/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/89/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/90/</loc>
-    
-  </url>
-  
-  <url>
-    <loc>https://18f.gsa.gov/blog/page/91/</loc>
-    
-  </url>
-</urlset>
-`;
-
-export { source, testRobotsTxt, testSitemapXml };
