+++ nest-cli.json
@@ -107,15 +107,6 @@
         "tsConfigPath": "apps/cli/tsconfig.app.json"
       }
     },
-    "solutions-scanner": {
-      "type": "library",
-      "root": "libs/solutions-scanner",
-      "entryFile": "index",
-      "sourceRoot": "libs/solutions-scanner/src",
-      "compilerOptions": {
-        "tsConfigPath": "libs/solutions-scanner/tsconfig.lib.json"
-      }
-    },
     "storage": {
       "type": "library",
       "root": "libs/storage",
